+++ backend/src/app/api/v1/endpoints/query.py
@@ -7,14 +7,16 @@ from fastapi import APIRouter, Depends, HTTPException
 
 from app.core.dependencies import get_llm_service, get_vector_db_service
 from app.schemas.query_api import (
+    QueryAnswer,
+    QueryAnswerResponse,
     QueryRequestSchema,
-    QueryResponseSchema,
     QueryResult,
 )
-from app.services.llm.base import LLMService
+from app.services.llm.base import CompletionService
 from app.services.query_service import (
     decomposition_query,
     hybrid_query,
+    inference_query,
     simple_vector_query,
 )
 from app.services.vector_db.base import VectorDBService
@@ -26,12 +28,12 @@ router = APIRouter(tags=["query"])
 logger.info("Query router initialized")
 
 
-@router.post("", response_model=QueryResponseSchema)
+@router.post("", response_model=QueryAnswerResponse)
 async def run_query(
     request: QueryRequestSchema,
-    llm_service: LLMService = Depends(get_llm_service),
+    llm_service: CompletionService = Depends(get_llm_service),
     vector_db_service: VectorDBService = Depends(get_vector_db_service),
-) -> QueryResponseSchema:
+) -> QueryAnswerResponse:
     """
     Run a query and generate a response.
 
@@ -43,7 +45,7 @@ async def run_query(
     ----------
     request : QueryRequestSchema
         The incoming query request.
-    llm_service : LLMService
+    llm_service : CompletionService
         The language model service.
     vector_db_service : VectorDBService
         The vector database service.
@@ -58,6 +60,30 @@ async def run_query(
     HTTPException
         If there's an error processing the query.
     """
+    if request.document_id == "00000000000000000000000000000000":
+        query_response = await inference_query(
+            request.prompt.query,
+            request.prompt.rules,
+            request.prompt.type,
+            llm_service,
+        )
+
+        if not isinstance(query_response, QueryResult):
+            query_response = QueryResult(**query_response)
+
+        answer = QueryAnswer(
+            id=uuid.uuid4().hex,
+            document_id=request.document_id,
+            prompt_id=request.prompt.id,
+            answer=query_response.answer,
+            type=request.prompt.type,
+        )
+        response_data = QueryAnswerResponse(
+            answer=answer, chunks=query_response.chunks
+        )
+
+        return response_data
+
     try:
         logger.info(f"Received query request: {request.model_dump()}")
 
@@ -86,13 +112,24 @@ async def run_query(
         if not isinstance(query_response, QueryResult):
             query_response = QueryResult(**query_response)
 
-        response_data = QueryResponseSchema(
-            id=str(uuid.uuid4()),
+        # response_data = QueryResponseSchema(
+        #     id=str(uuid.uuid4()),
+        #     document_id=request.document_id,
+        #     prompt_id=request.prompt.id,
+        #     type=request.prompt.type,
+        #     answer=query_response.answer,
+        #     chunks=query_response.chunks,
+        # )
+
+        answer = QueryAnswer(
+            id=uuid.uuid4().hex,
             document_id=request.document_id,
             prompt_id=request.prompt.id,
-            type=request.prompt.type,
             answer=query_response.answer,
-            chunks=query_response.chunks,
+            type=request.prompt.type,
+        )
+        response_data = QueryAnswerResponse(
+            answer=answer, chunks=query_response.chunks
         )
 
         return response_data