+++ backend/src/app/services/query_service.py
@@ -5,7 +5,11 @@ from typing import Any, Awaitable, Callable, List
 
 from app.models.query_core import Chunk, FormatType, QueryType, Rule
 from app.schemas.query_api import QueryResult, SearchResponse
-from app.services.llm_service import LLMService, generate_response
+from app.services.llm_service import (
+    CompletionService,
+    generate_inferred_response,
+    generate_response,
+)
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -40,7 +44,7 @@ async def process_query(
     document_id: str,
     rules: List[Rule],
     format: FormatType,
-    llm_service: LLMService,
+    llm_service: CompletionService,
     vector_db_service: Any,
 ) -> QueryResult:
     """Process the query based on the specified type."""
@@ -71,7 +75,7 @@ async def decomposition_query(
     document_id: str,
     rules: List[Rule],
     format: FormatType,
-    llm_service: LLMService,
+    llm_service: CompletionService,
     vector_db_service: Any,
 ) -> QueryResult:
     """Process the query based on the decomposition type."""
@@ -91,7 +95,7 @@ async def hybrid_query(
     document_id: str,
     rules: List[Rule],
     format: FormatType,
-    llm_service: LLMService,
+    llm_service: CompletionService,
     vector_db_service: Any,
 ) -> QueryResult:
     """Process the query based on the hybrid type."""
@@ -111,7 +115,7 @@ async def simple_vector_query(
     document_id: str,
     rules: List[Rule],
     format: FormatType,
-    llm_service: LLMService,
+    llm_service: CompletionService,
     vector_db_service: Any,
 ) -> QueryResult:
     """Process the query based on the simple vector type."""
@@ -124,3 +128,21 @@ async def simple_vector_query(
         llm_service,
         vector_db_service,
     )
+
+
+async def inference_query(
+    query: str,
+    rules: List[Rule],
+    format: FormatType,
+    llm_service: CompletionService,
+) -> QueryResult:
+    """Generate a response, no need for vector retrieval."""
+    # Since we are just answering this query based on data provided in the query,
+    # ther is no need to retrieve any chunks from the vector database.
+
+    answer = await generate_inferred_response(
+        llm_service, query, rules, format
+    )
+    answer_value = answer["answer"]
+
+    return QueryResult(answer=answer_value, chunks=[])